Split NativeWindow.WindowClass into separate file
